requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
psycopg==3.2.3
pyasn1==0.6.1
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.24.0.post1
razorpay